+++ src/preprocessing/mod.rs
@@ -1,10 +1,12 @@
 //! Fitted dense preprocessing transformers.
 
 mod binarizer;
+mod expansion;
 mod function_transformer;
 mod max_abs_scaler;
 mod min_max_scaler;
 mod normalizer;
+mod polynomial_features;
 mod robust_scaler;
 mod scaling;
 mod standard_scaler;
@@ -14,5 +16,6 @@ pub use function_transformer::{ElementwiseFn, FunctionTransformer, FunctionTrans
 pub use max_abs_scaler::{MaxAbsScaler, MaxAbsScalerParams};
 pub use min_max_scaler::{MinMaxScaler, MinMaxScalerParams};
 pub use normalizer::{Norm, Normalizer, NormalizerParams};
+pub use polynomial_features::{PolynomialFeatures, PolynomialFeaturesParams};
 pub use robust_scaler::{RobustScaler, RobustScalerParams};
 pub use standard_scaler::{StandardScaler, StandardScalerParams};